refactor: move solana client under wallet
